requirements.txt
huggingface_hub
langchain
streamlit
gradio
python-dotenv
pyPDF2
transformers
accelerate
bitsandbytes

-e .